requirements.txt
asgiref==3.3.4
Django==3.1.8
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.2
pytz==2021.1
sqlparse==0.4.1
